todos fixed hundo p
