+++ src/Components/Productivity/Productivity.css
@@ -12,6 +12,7 @@
   -webkit-box-shadow: 4px 0px 5px 0px rgba(0, 0, 0, 0.59);
   -moz-box-shadow: 4px 0px 5px 0px rgba(0, 0, 0, 0.59);
   z-index: 10;
+  overflow-y: scroll;
 }
 
 .pomodoro-section {
